refactored out the grammar handling. handles epsilons better
